Schedule Form Update
Note: I haven't renamed the actual variable from Schedual to Schedule yet to avoid possible compatibility issues.
This is an update to the Schedule Form that has the ability to save/load from a separate .dat file that will be automatically created if it doesn't already exist. I also added print functionality to this form that I had gotten online, though I haven't tested if it actually would print yet.